Add ktlint and refactor code accordng analysis result
