+++ build.gradle.kts
@@ -20,10 +20,30 @@ buildscript {
 
 plugins {
     id(gradlePlugins.Plugins.GRADLE_VERSIONS_PLUGIN) version gradlePlugins.PluginVersions.GRADLE_VER_PLUGIN_VERSION
+    id(gradlePlugins.Plugins.KTLINT) version gradlePlugins.PluginVersions.KTLINT_VERSION
 }
 
 allprojects {
     repositories.applyDefault()
+}
+
+subprojects {
+    apply {
+        plugin(gradlePlugins.Plugins.KTLINT)
+    }
+
+    ktlint {
+        debug.set(false)
+        verbose.set(true)
+        android.set(false)
+        outputToConsole.set(true)
+        ignoreFailures.set(false)
+        enableExperimentalRules.set(true)
+        filter {
+            exclude("**/generated/**")
+            include("**/kotlin/**")
+        }
+    }
 }
 
 tasks.register("clean", Delete::class) {
