+++ buildSrc/src/main/kotlin/AndroidConfig.kt
@@ -7,7 +7,7 @@ object AndroidConfig {
     const val COMPILE_SDK_VERSION = 30
     const val BUILD_TOOLS_VERSION = "30.0.0"
 
-    const val APPLICATION_ID = "com.danteyu.studio.my_android_template"
+    const val APPLICATION_ID = "com.danteyu.studio.template"
     const val MIN_SDK_VERSION = 24
     const val TARGET_SDK_VERSION = 30
