+++ buildSrc/src/main/kotlin/gradlePlugins/PluginVersions.kt
@@ -3,4 +3,5 @@ package gradlePlugins
 
 object PluginVersions {
     const val GRADLE_VER_PLUGIN_VERSION = "0.38.0"
+    const val KTLINT_VERSION = "10.0.0"
 }
