+++ buildSrc/src/main/kotlin/gradlePlugins/Plugins.kt
@@ -19,4 +19,5 @@ object Plugins {
     const val NAV_SAFEARGS = "androidx.navigation.safeargs.kotlin"
 
     const val GRADLE_VERSIONS_PLUGIN = "com.github.ben-manes.versions"
+    const val KTLINT = "org.jlleitschuh.gradle.ktlint"
 }
