+++ library/src/main/java/com/danteyu/studio/library/MyClass.kt
@@ -1,4 +1,3 @@
 package com.danteyu.studio.library
 
-class MyClass {
-}
+class MyClass
